requirements.txt
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==12.1.1
pyserial==3.5
ultralytics==8.2.3
ultralytics==8.4.30
